learning opencv

+++ chapter3/managers.py
@@ -21,7 +21,8 @@ class CaptureManager(object):
         self._videoWriter = None
         
         self._startTime = None
-        self._framesElapsed = long(0)
+        # self._framesElapsed = long(0) python3 remove long
+        self._framesElapsed = 0
         self._fpsEstimate = None
     
     @property
